Rabbitmq update

+++ src/main/java/ru/itfbgroup/notificationservice/service/NotificationService.java
@@ -53,6 +53,6 @@ public class NotificationService {
 
         }
         mailSender.send(msg);
-        file.delete();
+        if (file != null) file.delete();
     }
 }
